refactor: add more exceptions
